исправлено отрицаиние и тест для него

+++ include/tbitfield.h
@@ -5,7 +5,7 @@
 
 #include <iostream>
 using namespace std;
-typedef unsigned int TELEM;
+typedef unsigned short TELEM;
 
 class TBitField
 {
